Link Devfile 2.1.0 to 2.0.0.   Devfile 2.1.0  enhances 2.0.0 and they  do not have conflicts.    2.1.0/2.0.0 to share schema and struct.
