+++ pkg/devfile/validate/validate.go
@@ -8,7 +8,6 @@ import (
 
 	v100 "github.com/openshift/odo/pkg/devfile/parser/data/1.0.0"
 	v200 "github.com/openshift/odo/pkg/devfile/parser/data/2.0.0"
-	v210 "github.com/openshift/odo/pkg/devfile/parser/data/2.1.0"
 )
 
 // ValidateDevfileData validates whether sections of devfile are odo compatible
@@ -24,11 +23,6 @@ func ValidateDevfileData(data interface{}) error {
 
 	if typeData == reflect.TypeOf(&v200.Devfile200{}) {
 		d := data.(*v200.Devfile200)
-		components = d.GetComponents()
-	}
-
-	if typeData == reflect.TypeOf(&v210.Devfile210{}) {
-		d := data.(*v210.Devfile210)
 		components = d.GetComponents()
 	}
