+++ tests/examples/source/devfiles/nodejs/devfile_deploy.yaml
@@ -1,8 +1,7 @@
-schemaVersion: 2.0.0
+schemaVersion: 2.1.0
 metadata:
   name: nodejs
   version: 1.0.0
-  alpha.build-dockerfile: "https://raw.githubusercontent.com/groeges/devfile-registry/master/devfiles/nodejs/Dockerfile"
   alpha.deployment-manifest: "https://raw.githubusercontent.com/neeraj-laad/nodejs-stack-registry/build-deploy/devfiles/nodejs-basic/deploy/k8s-deploy.yaml"
 projects:
   - name: express
@@ -27,6 +26,13 @@ components:
             protocol: tcp
             scheme: http
             type: terminal
+  - dockerfile:
+      name: dockerfile-build
+      source: 
+         sourceDir: "src"
+         location: "https://github.com/ranakan19/golang-ex.git"
+      dockerfilePath: "https://raw.githubusercontent.com/wtam2018/test/master/nodejs-dockerfiile"
+      destination: 
 commands:
   - exec:
       id: download app dependencies
