+++ tests/examples/source/devfilesV2/nodejs/devfile.yaml
@@ -1,7 +1,6 @@
-schemaVersion: "2.0.0"
+schemaVersion: "2.1.0"
 metadata:
   name: test-devfile
-  alpha.build-dockerfile: "https://raw.githubusercontent.com/neeraj-laad/nodejs-stack-registry/build-deploy/devfiles/nodejs-basic/build/Dockerfile"
   alpha.deployment-manifest: "https://raw.githubusercontent.com/groeges/devfile-registry/master/devfiles/nodejs/deploy_deployment.yaml"
 projects:
   - name: nodejs-web-app
@@ -24,6 +23,13 @@ components:
             protocol: tcp
             scheme: http
             type: terminal
+  - dockerfile:
+      name: dockerfile-build
+      source: 
+         sourceDir: "src"
+         location: "https://github.com/ranakan19/golang-ex.git"
+      dockerfilePath: "https://raw.githubusercontent.com/wtam2018/test/master/nodejs-dockerfiile"
+      destination: 
 commands:
   - exec:
       id: download dependencies
