+++ tests/integration/devfile/cmd_devfile_deploy_test.go
@@ -48,7 +48,7 @@ var _ = Describe("odo devfile deploy command tests", func() {
 			helper.CmdShouldPass("odo", "create", "nodejs", "--project", namespace, cmpName)
 			helper.CmdShouldPass("odo", "url", "create", "--port", "3000")
 
-			helper.CopyExampleDevFile(filepath.Join("source", "devfilesV2.1.0", "nodejs", "devfile.yaml"), filepath.Join(context, "devfile.yaml"))
+			helper.CopyExampleDevFile(filepath.Join("source", "devfilesV2", "nodejs", "devfile.yaml"), filepath.Join(context, "devfile.yaml"))
 			output := helper.CmdShouldPass("odo", "deploy", "--tag", imageTag)
 			cliRunner.WaitAndCheckForExistence("buildconfig", namespace, 1)
 			Expect(output).NotTo(ContainSubstring("does not point to a valid Dockerfile"))
@@ -61,7 +61,7 @@ var _ = Describe("odo devfile deploy command tests", func() {
 		It("Should error out with 'URL does not point to a valid Dockerfile'", func() {
 			helper.CmdShouldPass("odo", "create", "nodejs", "--project", namespace, cmpName)
 			helper.CmdShouldPass("odo", "url", "create", "--port", "3000")
-			helper.CopyExampleDevFile(filepath.Join("source", "devfilesV2.1.0", "nodejs", "devfile.yaml"), filepath.Join(context, "devfile.yaml"))
+			helper.CopyExampleDevFile(filepath.Join("source", "devfilesV2", "nodejs", "devfile.yaml"), filepath.Join(context, "devfile.yaml"))
 
 			err := helper.ReplaceDevfileField("devfile.yaml", "dockerfilePath", "https://google.com")
 			Expect(err).To(BeNil())
